+++ CHANGES.md
@@ -1,5 +1,12 @@
 [unreleased]
 
+#### 3.3.1 / 2023-03-17 ☘️
+* update readme
+* sort listed plugins in 'Report a Bug'
+* add mu-plugins in 'Report a Bug'
+* fix for multisite
+* initiate plugin in `plugins_loaded`
+
 #### 3.3.0 / 2023-03-16
 * added `Report a Bug` feature, thanks @costdev, @ironprogrammer